fix: rename AssertError to InvalidOrder
