+++ src/index.ts
@@ -1,7 +1,7 @@
 import { AssertOrder } from './AssertOrder'
 export default AssertOrder
 
-export * from './AssertError'
 export * from './AssertOrder'
 export * from './interfaces'
+export * from './InvalidOrder'
 export * from './satisfy'
